+++ lib/controllers/session_controller.dart
@@ -1,22 +1,28 @@
 import 'dart:async';
 import 'dart:convert';
+import 'dart:io';
 
 import 'package:agora_rtc_engine/rtc_engine.dart';
+import 'package:agora_rtm/agora_rtm.dart';
 import 'package:agora_uikit/models/agora_channel_data.dart';
 import 'package:agora_uikit/models/agora_connection_data.dart';
 import 'package:agora_uikit/models/agora_event_handlers.dart';
 import 'package:agora_uikit/models/agora_settings.dart';
 import 'package:agora_uikit/models/agora_user.dart';
+import 'package:agora_uikit/models/rtm_message.dart';
 import 'package:agora_uikit/src/enums.dart';
 import 'package:flutter/material.dart';
 import 'package:http/http.dart' as http;
 import 'package:permission_handler/permission_handler.dart';
+import '../src/enums.dart';
 
 class SessionController extends ValueNotifier<AgoraSettings> {
   SessionController()
       : super(
           AgoraSettings(
             engine: null,
+            agoraRtmChannel: null,
+            agoraRtmClient: null,
             users: [],
             mainAgoraUser: AgoraUser(
               uid: 0,
@@ -31,10 +37,75 @@ class SessionController extends ValueNotifier<AgoraSettings> {
             clientRole: ClientRole.Broadcaster,
             localUid: 0,
             generatedToken: null,
+            generatedRtmId: null,
             layoutType: Layout.grid,
+            displaySnackbar: false,
+            muteRequest: MicState.unmuted,
+            cameraRequest: CameraState.enabled,
+            showMicMessage: false,
+            showCameraMessage: false,
           ),
         );
 
+  /// Function to initialize the Agora RTM client.
+  Future<void> initializeRtm() async {
+    value = value.copyWith(
+      agoraRtmClient: await AgoraRtmClient.createInstance(
+        value.connectionData!.appId,
+      ),
+    );
+
+    value.agoraRtmClient?.onMessageReceived =
+        (AgoraRtmMessage message, String peerId) {
+      print('Peer msg : $message, from : $peerId');
+      Message msg = Message(text: message.text);
+
+      String? messageType;
+      msg.toJson().forEach((key, val) {
+        if (key == "text") {
+          var json = jsonDecode(val);
+          messageType = json['messageType'];
+        }
+      });
+      _onMessageReceived(messageType: messageType!, message: msg.toJson());
+    };
+
+    value.agoraRtmClient?.onConnectionStateChanged = (int state, int reason) {
+      print(
+          'Connection state changed : ${state.toString()}, reason : ${reason.toString()}');
+      if (state == 5) {
+        value.agoraRtmClient?.logout();
+      }
+    };
+
+    value.agoraRtmClient?.onLocalInvitationReceivedByPeer =
+        (AgoraRtmLocalInvitation invitation) {
+      print(
+          'Local invitation received by peer : ${invitation.calleeId}, content : ${invitation.content}');
+    };
+
+    value.agoraRtmClient?.onRemoteInvitationReceivedByPeer =
+        (AgoraRtmRemoteInvitation invitation) {
+      print(
+          'Remote invitation received by peer : ${invitation.callerId}, content : ${invitation.content}');
+    };
+
+    value.agoraRtmClient?.onError = () {
+      print('Error Occurred');
+    };
+  }
+
+  /// Function to join the RTM channel and send the user data to everyone inside that channel.
+  Future<void> rtmMethods() async {
+    await _loginToRtm();
+    await _joinRtmChannel();
+    await _sendUserData(
+      toChannel: true,
+      username: value.connectionData!.username!,
+    );
+  }
+
+  /// Function to initialize the Agora RTC engine.
   Future<void> initializeEngine(
       {required AgoraConnectionData agoraConnectionData}) async {
     value = value.copyWith(
@@ -48,6 +119,11 @@ class SessionController extends ValueNotifier<AgoraSettings> {
     );
   }
 
+  Future<void> askForUserCameraAndMicPermission() async {
+    await [Permission.camera, Permission.microphone].request();
+  }
+
+  /// Function to trigger all the AgoraEventHandlers.
   void createEvents(AgoraEventHandlers agoraEventHandlers) async {
     value.engine?.setEventHandler(
       RtcEngineEventHandler(
@@ -238,12 +314,13 @@ class SessionController extends ValueNotifier<AgoraSettings> {
               clientRole: value.clientRole,
             ),
           );
-
+          if (value.connectionData!.username != null) {
+            rtmMethods();
+          }
           agoraEventHandlers.joinChannelSuccess?.call(channel, uid, elapsed);
         },
         leaveChannel: (stats) {
           _clearUsers();
-
           agoraEventHandlers.leaveChannel?.call(stats);
         },
         userJoined: (uid, elapsed) {
@@ -391,10 +468,16 @@ class SessionController extends ValueNotifier<AgoraSettings> {
         agoraChannelData.audioProfile, agoraChannelData.audioScenario);
   }
 
+  /// Function to join the video call.
   Future<void> joinVideoChannel() async {
+    // [generatedRtmId] is the unique ID for a user generated using the timestamp in milliseconds.
+    value = value.copyWith(
+      generatedRtmId: value.connectionData!.rtmUid ??
+          DateTime.now().millisecondsSinceEpoch.toString(),
+    );
     await value.engine?.enableVideo();
     await value.engine?.enableAudioVolumeIndication(200, 3, true);
-    if (value.connectionData!.tokenUrl != null) {
+    if (value.connectionData?.tokenUrl != null) {
       await _getToken(
         tokenUrl: value.connectionData!.tokenUrl,
         channelName: value.connectionData!.channelName,
@@ -457,8 +540,13 @@ class SessionController extends ValueNotifier<AgoraSettings> {
     await value.engine?.switchCamera();
   }
 
+  /// Function to dispose the RTC and RTM engine.
   Future<void> endCall() async {
     await value.engine?.leaveChannel();
+    if (value.connectionData!.username != null) {
+      await value.agoraRtmChannel?.leave();
+      await value.agoraRtmClient?.logout();
+    }
     await value.engine?.destroy();
     // dispose();
   }
@@ -537,6 +625,9 @@ class SessionController extends ValueNotifier<AgoraSettings> {
     String? channelName,
     int uid = 0,
   }) async {
+    if (value.connectionData!.username != null) {
+      await _getRtmToken(tokenUrl: tokenUrl);
+    }
     final response = await http
         .get(Uri.parse('$tokenUrl/rtc/$channelName/publisher/uid/$uid'));
     if (response.statusCode == 200) {
@@ -548,7 +639,295 @@ class SessionController extends ValueNotifier<AgoraSettings> {
     }
   }
 
+  Future<void> _getRtmToken({String? tokenUrl}) async {
+    final String url = "$tokenUrl/rtm/${value.generatedRtmId}";
+    final rtmResponse = await http.get(Uri.parse(url));
+    print(url);
+    if (rtmResponse.statusCode == 200) {
+      value = value.copyWith(
+        generatedRtmToken: jsonDecode(rtmResponse.body)['rtmToken'],
+      );
+    } else {
+      print(rtmResponse.reasonPhrase);
+      print('Failed to generate the rtm token : ${rtmResponse.statusCode}');
+    }
+  }
+
   void updateLayoutType({required Layout updatedLayout}) {
     value = value.copyWith(layoutType: updatedLayout);
+  }
+
+  Future<void> _loginToRtm() async {
+    if (!value.isLoggedIn) {
+      try {
+        await value.agoraRtmClient?.login(
+          value.connectionData!.tempRtmToken ?? value.generatedRtmToken,
+          value.generatedRtmId!,
+        );
+        value = value.copyWith(isLoggedIn: true);
+        print(
+            'Username : ${value.connectionData!.username} and rtmId : ${value.generatedRtmId} logged in');
+      } catch (e) {
+        print('Error occurred while trying to login. ${e.toString()}');
+      }
+    }
+  }
+
+  Future<AgoraRtmChannel?> _createChannel(
+      {required String rtmChannelName}) async {
+    AgoraRtmChannel? channel =
+        await value.agoraRtmClient?.createChannel(rtmChannelName);
+
+    if (channel != null) {
+      channel.onMessageReceived =
+          (AgoraRtmMessage message, AgoraRtmMember member) {
+        print('Channel msg : ${message.text}, from : ${member.userId}');
+        Message msg = Message(text: message.text);
+        _onMessageReceived(messageType: "UserData", message: msg.toJson());
+      };
+
+      channel.onMemberJoined = (AgoraRtmMember member) {
+        print('Member joined : ${member.userId}');
+        _sendUserData(
+          toChannel: false,
+          username: value.connectionData!.username!,
+          peerRtmId: member.userId,
+        );
+      };
+
+      channel.onMemberLeft = (AgoraRtmMember member) {
+        print('Member left : ${member.userId}');
+
+        if (value.userRtmMap!.containsKey(member.userId)) {
+          _removeFromUserRtmMap(rtmId: member.userId);
+        }
+
+        if (value.uidToUserIdMap!.containsValue(member.userId)) {
+          for (var i = 0; i < value.uidToUserIdMap!.length; i++) {
+            int rtcId = value.uidToUserIdMap!.keys.elementAt(i);
+            _removeFromUidToUserMap(rtcId: rtcId);
+          }
+        }
+      };
+
+      channel.onMemberCountUpdated = (int count) {
+        print('Member count updated : $count');
+      };
+    }
+    return channel;
+  }
+
+  Future<void> _joinRtmChannel() async {
+    if (!value.isInChannel) {
+      try {
+        value = value.copyWith(
+          agoraRtmChannel: await _createChannel(
+              rtmChannelName: value.connectionData?.rtmChannelName ??
+                  value.connectionData!.channelName),
+        );
+        await value.agoraRtmChannel?.join();
+        value = value.copyWith(isInChannel: true);
+      } catch (e) {
+        print('RTM Join channel error : ${e.toString()}');
+      }
+    }
+  }
+
+  Future<void> _sendUserData({
+    required bool toChannel,
+    required String username,
+    String? peerRtmId,
+  }) async {
+    String platform = Platform.isAndroid ? "android" : "ios";
+    int ts = DateTime.now().millisecondsSinceEpoch;
+
+    String json = '''{
+      "messageType": "UserData",
+      "rtmId": "${value.generatedRtmId}",
+      "rtcId": ${value.localUid},
+      "username": "$username",
+      "role": "${value.clientRole}",
+      "agora": {
+        "rtm": "1.0.1",
+        "rtc": "4.0.6"
+      },
+      "uikit": {
+        "platform": "$platform",
+        "framework": "flutter",
+        "version": "1.0.0"
+      }
+    }''';
+
+    Message message = Message(text: json, ts: ts, offline: false);
+    AgoraRtmMessage msg = AgoraRtmMessage.fromJson(message.toJson());
+
+    if (value.agoraRtmChannel != null && toChannel) {
+      await value.agoraRtmChannel?.sendMessage(msg);
+      print('User data sent to channel');
+    } else if (value.agoraRtmClient != null &&
+        !toChannel &&
+        peerRtmId != null) {
+      await value.agoraRtmClient?.sendMessageToPeer(peerRtmId, msg);
+      print('User data sent to peer');
+    } else {
+      print("No user in the channel");
+    }
+  }
+
+  void askForUserMic({required int index, required bool isMicEnabled}) {
+    String? peerId;
+    String json = '''{
+      "messageType": "MuteRequest",
+      "rtcId": ${value.users[index].uid},
+      "mute": $isMicEnabled,
+      "device": "0", 
+      "isForceFul": "false"
+    }''';
+
+    Message message = Message(text: json);
+    AgoraRtmMessage msg = AgoraRtmMessage.fromJson(message.toJson());
+    value.uidToUserIdMap!.forEach((key, val) {
+      if (key == value.users[index].uid) {
+        peerId = val;
+        print("PEER RTM ID : $peerId");
+        if (value.isLoggedIn) {
+          value.agoraRtmClient?.sendMessageToPeer(peerId!, msg);
+        } else {
+          print("USER NOT LOGGED IN (MIC)");
+        }
+      } else {
+        print("Peer RTM ID not found");
+      }
+    });
+  }
+
+  void askForUserCamera({required int index, required bool isCameraEnabled}) {
+    String? peerId;
+    String json = '''{
+      "messageType": "CameraRequest",
+      "rtcId": ${value.users[index].uid},
+      "mute": $isCameraEnabled,
+      "device": "1",
+      "isForceFul": "false"
+    }''';
+
+    Message message = Message(text: json);
+    AgoraRtmMessage msg = AgoraRtmMessage.fromJson(message.toJson());
+    value.uidToUserIdMap!.forEach((key, val) {
+      if (key == value.users[index].uid) {
+        peerId = val;
+        value.agoraRtmClient?.sendMessageToPeer(peerId!, msg);
+      } else {
+        print("Peer RTM ID not found");
+      }
+    });
+  }
+
+  void _addToUidUserMap({required int rtcId, required String rtmId}) {
+    Map<int, String> tempMap = {};
+    tempMap.addAll(value.uidToUserIdMap ?? {});
+    if (rtcId != 0) {
+      tempMap.putIfAbsent(rtcId, () => rtmId);
+    }
+    value = value.copyWith(uidToUserIdMap: tempMap);
+  }
+
+  void _addToUserRtmMap(
+      {required String rtmId, required Map<String, dynamic> message}) {
+    Map<String, Map<String, dynamic>> tempMap = {};
+    tempMap.addAll(value.userRtmMap ?? {});
+    tempMap.putIfAbsent(rtmId, () => message);
+    value = value.copyWith(userRtmMap: tempMap);
+  }
+
+  void _removeFromUidToUserMap({required int rtcId}) {
+    Map<int, String> tempMap = {};
+    tempMap = value.uidToUserIdMap!;
+    for (int i = 0; i < tempMap.length; i++) {
+      if (tempMap.keys.elementAt(i) == rtcId) {
+        tempMap.remove(rtcId);
+      }
+    }
+    value = value.copyWith(uidToUserIdMap: tempMap);
+  }
+
+  void _removeFromUserRtmMap({required String rtmId}) {
+    Map<String, Map<String, dynamic>> tempMap = {};
+    tempMap = value.userRtmMap!;
+    for (int i = 0; i < tempMap.length; i++) {
+      if (tempMap.keys.elementAt(i) == rtmId) {
+        tempMap.remove(rtmId);
+      }
+    }
+    value = value.copyWith(userRtmMap: tempMap);
+  }
+
+  void _onMessageReceived(
+      {required String messageType, required Map<String, dynamic> message}) {
+    switch (messageType) {
+      case "UserData":
+        print('User data received: $message');
+        message.forEach((key, val) {
+          if (key == "text") {
+            var json = jsonDecode(val);
+            String rtmId = json['rtmId'];
+            int rtcId = json['rtcId'];
+            print("RTM ID: $rtmId");
+            print("RTC ID : $rtcId");
+            _addToUidUserMap(rtcId: rtcId, rtmId: rtmId);
+            _addToUserRtmMap(rtmId: rtmId, message: message);
+          }
+        });
+        break;
+      case "MuteRequest":
+        int? rtcId;
+        bool? muted;
+        message.forEach((key, val) {
+          if (key == "text") {
+            var json = jsonDecode(val);
+            rtcId = json['rtcId'];
+            muted = json['mute'];
+            print("RTC ID: $rtcId");
+          }
+        });
+        value = value.copyWith(
+          displaySnackbar: true,
+          muteRequest: muted! ? MicState.unmuted : MicState.muted,
+          showCameraMessage: false,
+          showMicMessage: true,
+        );
+        Future.delayed(Duration(seconds: 10), () {
+          value = value.copyWith(
+            displaySnackbar: false,
+            showMicMessage: false,
+          );
+        });
+        break;
+      case "CameraRequest":
+        int? rtcId;
+        bool? disabled;
+        message.forEach((key, val) {
+          if (key == "text") {
+            var json = jsonDecode(val);
+            rtcId = json['rtcId'];
+            disabled = json['mute'];
+            print("RTC ID: $rtcId");
+          }
+        });
+        value = value.copyWith(
+          displaySnackbar: true,
+          cameraRequest: disabled! ? CameraState.enabled : CameraState.disabled,
+          showMicMessage: false,
+          showCameraMessage: true,
+        );
+        Future.delayed(Duration(seconds: 10), () {
+          value = value.copyWith(
+            displaySnackbar: false,
+            showCameraMessage: false,
+          );
+        });
+        break;
+      default:
+    }
   }
 }