+++ lib/src/enums.dart
@@ -1,2 +1,4 @@
 enum Layout { grid, floating }
 enum BuiltInButtons { callEnd, switchCamera, toggleCamera, toggleMic }
+enum MicState { muted, unmuted }
+enum CameraState { enabled, disabled }